Fix for FHIR-35566
FHIR-335566: Replaced local value set with VSAC equivalent (http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113762.1.4.1115.9).
Note that the direct link to the VSAC valueset as requested in the JIRA ticket will require a UMLS account. Otherwise, the link will fail. The FHIR IG Publisher expansion display works in the IG if we create our own value set, which then references the VSAC value set. That however is one extra level of indirection and not exactly what was requested in the JIRA ticket.