Fix for FHIR-35567
Applied fix notes:
Based on the more updated ballot resolution description, I'm assuming that the original comment from Brian Meshell in March was superseded by the ballot resolution and vote on May 4, 2023.
Also, I'm replacing the value set name from PADIInterventionPreferencesNarrative to PADIInterventionPreferencessEndOfLifeGrouping, which is the actual value set name in VSAC for that OID of 2.16.840.1.113762.1.4.1115.9.