Update content_type_overview.md
Minor editorial changes for consistency and to mention establishing a HCA with the Person-authored document type.
FHIR-48858: Added guidance to clarify the role of a facilitator in the creation of advance healthcare directive documents. 
FHIR-43185.  explain the Mental Health Advance Directive document and how the permitted actions for Healthcare Agents in these types of documents may be tailored to the types of decisions and powers for a Mental Health focused healthcare agent.